Add files via upload
Game made in Scratch, In this Game there are one spaceship and  to control that spaceship press the arrow keys, and make spaceship travel to the planet which will appear after few seconds and when the spaceship travels into space, spaceship avoid to touch the meteorites which are moving on screen, if spaceship touches the meteorites then spaceship will disappear for some seconds and appear again. when the spaceship touches the planet the game will be over and at the last gives player a message that how much time player taken for reaching the planet and display button for restart the game.